Write a program that prints the first 50 Fibonacci numbers, starting with 1 and 2, followed by a new line

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -7,19 +7,19 @@
 int main(void)
 {
 	int i = 0;
-	int n = 1, m = 2;
+	long int n = 1, m = 2;
 
 	while (i < 50)
 	{
 		if (i == 0)
-			printf("%d", n);
+			printf("%ld", n);
 		else if (i == 1)
-			printf(", %d", m);
+			printf(", %ld", m);
 		else
 		{
 			m += n;
 			n = m - n;
-			printf(", %d", m);
+			printf(", %ld", m);
 		}
 		i++;
 	}
